refactor: extract AI detector to separate package
AI検出機能を独立したパッケージに分離
- packages/ai-detector を新規作成
- packages/detectors から ai-* ファイルを削除
- detectors パッケージから re-export で後方互換性を維持
