+++ packages/ai-detector/package.json
@@ -1,0 +1,10 @@
+{
+  "name": "@service-policy-auditor/ai-detector",
+  "version": "0.0.1",
+  "type": "module",
+  "main": "./src/index.ts",
+  "types": "./src/index.ts",
+  "exports": {
+    ".": "./src/index.ts"
+  }
+}
