+++ packages/ai-detector/src/index.ts
@@ -1,0 +1,29 @@
+/**
+ * @service-policy-auditor/ai-detector
+ *
+ * AIプロンプト検出パッケージ
+ * AIサービスへのリクエスト/レスポンスを検出・解析
+ */
+
+// Types
+export type {
+  InferredProvider,
+  AIDetectionMethod,
+  CapturedAIPrompt,
+  AIPromptContent,
+  AIResponseContent,
+  AIPromptSentDetails,
+  AIResponseReceivedDetails,
+  AIMonitorConfig,
+} from "./types.js";
+
+export { DEFAULT_AI_MONITOR_CONFIG } from "./types.js";
+
+// Detector
+export {
+  isAIRequestBody,
+  extractPromptContent,
+  extractModel,
+  extractResponseContent,
+  inferProviderFromResponse,
+} from "./detector.js";
