+++ packages/ai-detector/src/types.ts
