+++ packages/detectors/package.json
@@ -10,6 +10,7 @@
     "./patterns": "./src/patterns.ts"
   },
   "dependencies": {
+    "@service-policy-auditor/ai-detector": "workspace:*",
     "@service-policy-auditor/csp": "workspace:*",
     "@service-policy-auditor/nrd": "workspace:*",
     "@service-policy-auditor/typosquat": "workspace:*"
